+++ benches/bench.rs
@@ -23,7 +23,7 @@ demux_context!(NullDemuxContext, NullFilterSwitch);
 impl NullDemuxContext {
     fn do_construct(&mut self, req: demultiplex::FilterRequest<'_, '_>) -> NullFilterSwitch {
         match req {
-            demultiplex::FilterRequest::ByPid(packet::Pid::PAT) => {
+            demultiplex::FilterRequest::ByPid(psi::pat::PAT_PID) => {
                 NullFilterSwitch::Pat(demultiplex::PatPacketFilter::default())
             }
             demultiplex::FilterRequest::ByPid(_) => {