+++ src/psi/mod.rs
@@ -798,7 +798,7 @@ mod test {
     impl NullDemuxContext {
         fn do_construct(&mut self, req: demultiplex::FilterRequest<'_, '_>) -> NullFilterSwitch {
             match req {
-                demultiplex::FilterRequest::ByPid(packet::Pid::PAT) => {
+                demultiplex::FilterRequest::ByPid(pat::PAT_PID) => {
                     NullFilterSwitch::Pat(demultiplex::PatPacketFilter::default())
                 }
                 demultiplex::FilterRequest::ByPid(_) => {
@@ -862,7 +862,7 @@ mod test {
         let state = Rc::new(RefCell::new(false));
         struct MockSectParse {
             state: Rc<RefCell<bool>>,
-        };
+        }
         impl WholeSectionSyntaxPayloadParser for MockSectParse {
             type Context = ();
             fn section<'a>(